+++ src/opensilex_python_client/variables/create.py
@@ -13,12 +13,13 @@ logger = get_logger(__name__)
 
 @dataclass
 class VariableData:
-    name: str
-    entity: str
-    characteristic: str
-    method: str
-    unit: str
-    datatype: str
+    uri: str | None = None
+    name: str = ""
+    entity: str = ""
+    characteristic: str = ""
+    method: str = ""
+    unit: str = ""
+    datatype: str = ""
     alternative_name: str = ""
     description: str = ""
     time_interval: str = ""
@@ -29,6 +30,7 @@ class VariableData:
 
     def to_dto(self) -> VariableCreationDTO:
         dto = VariableCreationDTO(
+            uri=self.uri,
             name=self.name,
             entity=self.entity,
             characteristic=self.characteristic,